new menu file and profile
