+++ angular.json
@@ -52,7 +52,7 @@
               "src/vendor/libs/ng-select/ng-select.scss",
               "src/vendor/styles/bootstrap.scss",
               "src/vendor/styles/appwork.scss",
-              "src/vendor/styles/theme-soft.scss",
+              "src/vendor/styles/theme-air.scss",
               "src/vendor/styles/colors.scss",
               "src/vendor/styles/uikit.scss"
             ],
