+++ src/app/+profile/profile.module.ts
@@ -1,16 +1,18 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { ProjectsModule } from '@app/+projects/projects.module';
 
 // **************************************************
 // Components & Services
 import { ProfileIndexComponent } from './+profile-index/profile-index.component';
 
 @NgModule({
-  declarations: [  
+  declarations: [
     ProfileIndexComponent,
   ],
   imports: [
-    CommonModule
+    CommonModule,
+    ProjectsModule
   ],
   providers: []
 })
