+++ src/app/+projects/+project-list/project-list.component.html
@@ -1,5 +1,3 @@
-<div class="row">
-  <div *ngFor="let project of projectsData" class="col-sm-6 col-xl-4 mb-4">
-    <app-project-card [project]="project"></app-project-card>
-  </div>
+<div class="col-xs-12 mb-3" *ngFor="let project of projectsData">
+  <app-project-card [project]="project"></app-project-card>
 </div>
