+++ src/app/app.routes.ts
@@ -2,19 +2,13 @@ import { AUTH_ROUTES } from './+auth/routes/auth.routes';
 import { Layout2Component } from './core/layout-2/layout-2.component';
 import { LayoutBlankComponent } from './core/layout-blank/layout-blank.component';
 import { USER_ROUTES } from './+users/routes/users.routes';
-import { HOME_ROUTES } from './+home/routes/home.routes';
 import { appRouteNames } from './app.routes.names';
 import { authRouteNames } from './+auth/routes/auth.routes.names';
 import { Projects_Routes } from './+projects/routes/projects.routes';
 import { PROFILE_ROUTES } from './+profile/routes/profile.routes';
 
 
-const MENU_ITEM_TYPES = {
-  HEADER: 'header',
-  ROUTE: 'route',
-  DIVIDER: 'divider',
-  SUBMENU: 'submenu'
-}
+
 
 export const APP_ROUTES: any = [
   {
@@ -22,7 +16,7 @@ export const APP_ROUTES: any = [
     component: Layout2Component,
     pathMatch: 'full',
     children: [
-      { path: '', children: HOME_ROUTES },
+      { path: '', children: PROFILE_ROUTES },
     ]
   },
   {
@@ -54,13 +48,3 @@ export const APP_ROUTES: any = [
     ]
   }
 ];
-
-export const APP_MENU = [
-  { type: MENU_ITEM_TYPES.HEADER, text: 'NAVIGATION', restricted: true },
-
-  // HOME
-  { type: MENU_ITEM_TYPES.ROUTE, text: 'Home', uri: '/', icon: 'fas fa-home', activeUriSegment: '/', restricted: false },
-
-  // USERS
-  { type: MENU_ITEM_TYPES.ROUTE, text: 'Users', icon: 'fas fa-user', uri: appRouteNames.USERS, activeUriSegment: appRouteNames.USERS, restricted: true}
-]
