+++ src/app/core/layout-navbar/layout-navbar.component.html
@@ -35,7 +35,7 @@
       <div ngbDropdown [placement]="isRTL ? 'bottom-left' : 'bottom-right'" class="demo-navbar-user nav-item">
         <a ngbDropdownToggle class="nav-link" href="javascript:void(0)">
           <span class="d-inline-flex align-items-center align-middle">
-            <img src="{{currentAuthUser?.photoURL}}" alt class="d-block ui-w-30 rounded-circle">
+            <!-- <img src="{{currentAuthUser?.photoURL}}" alt class="d-block ui-w-30 rounded-circle"> -->
             <span class="px-1 mr-lg-2 ml-2 ml-lg-0">{{currentAuthUser?.displayName}}</span>
           </span>
         </a>
