+++ src/app/core/layout-sidenav/layout-sidenav.component.html
@@ -1,4 +1,4 @@
-<sidenav [orientation]="orientation" [ngClass]="getClasses()">
+<sidenav [orientation]="orientation" [ngClass]="getClasses()" class="bg-white">
 
   <!-- Brand demo (see src/demo.css) -->
   <div class="app-brand brand-md text-center" *ngIf="orientation !== 'horizontal'">
